requirements.txt
aiomysql==0.3.2
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
atlastk==0.13.5
bcrypt==5.0.0
cffi==2.0.0
click==8.1.8
cloudbiolinux==0.1
cryptography==46.0.3
decorator==5.2.1
Deprecated==1.3.1
exceptiongroup==1.3.1
fabric==3.2.2
fastapi==0.124.0
h11==0.16.0
idna==3.11
invoke==2.2.1
iso8601==2.1.0
paramiko==4.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
PyNaCl==1.6.1
pypika-tortoise==0.6.3
pytz==2025.2
PyYAML==6.0.3
starlette==0.49.3
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
wrapt==2.0.1
